feat(DYLM-6): moved sendEmail to a hook
