+++ src/hooks/index.ts
@@ -1,0 +1,3 @@
+import { useSendEmail } from './useSendEmail';
+
+export { useSendEmail };
