+++ tsconfig.json
@@ -18,6 +18,7 @@
     "paths": {
       "@components/*": ["components/*"],
       "@context": ["context/index.tsx"],
+      "@hooks": ["hooks"],
       "@public/*": ["../public/assets/*"],
       "@services": ["services/index.ts"],
       "@styles/*": ["styles/*"],
